requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
bs4==0.0.1
chardet==3.0.4
distro==1.5.0
idna==2.9
multidict==4.7.5
numpy==1.18.3
pandas==1.0.3
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
soupsieve==2.0
tabula-py==2.1.0
tornado==6.0.4
yarl==1.4.2
